# swagger-template .yaml
swagger: "2.0"
info:
  description: "This is the API document for DKT-CN's `MessageHub` system's Template "
  version: "0.1"
  title: "DKT-CN MessageHub Template"
  termsOfService: Decathlon internal use only
  contact:
    email: "jerry.tan.partner@decathlon.com"
  license:
    name: Decathlon license
    url: "http://www.decathlon.fr/mentions-legales.html"
# host: ${api.metadata['api_host_proxy_intranet']}
# basePath: ${api.metadata['api_base_path_adapter']}
host: "messagehub.pp.dktapp.cloud"
basePath: "/api/adapter-fr"

tags:
  - name: "EventTypeManagement"
  - name: "TemplateManagement"

schemes:
  - "https"
paths:
  /v2/insert/event:
    post:
      tags:
        - EventTypeManagement
      summary: "api/v2/insert/event"
      description: "api/v2/insert/event"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: "The properties of communication."
          required: true
          schema:
            $ref: "#/definitions/eventTypeManagementRequest"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/eventTypeManagementResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'

  /v2/update/event/uuid:
    put:
      tags:
        - EventTypeManagement
      summary: "Preview the content of the message with given parameters."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ""
          required: true
          schema:
            $ref: "#/definitions/EventTypeManagementBodyPreview"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/eventTypeManagementResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'

  /v2/event/{uuid}:
    get:
      tags:
        - EventTypeManagement
      summary: "Get the sending status of a specific communication."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: uuid
          type: string
          description: "An UUID for this communication, generated by MessageHub system."
          required: true
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/eventTypeManagementResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'

  /v2/events:
    get:
      tags:
        - EventTypeManagement
      summary: "Get the sending history according to the given criterias."
      description: "`NOT RECOMMENDED TO USE`"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: eventName
          type: string
          description: "eventName"
        - in: query
          name: pageNumber
          type: string
          description: "pageNumber"
        - in: query
          name: pageSize
          type: string
          description: "pageSize"
        - in: query
          name: uuid
          type: string
          description: "uuid"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/eventTypeManagementResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'

  /v2/insert/template:
    post:
      tags:
        - TemplateManagement
      summary: "/api/email/message/insert-email-message-info."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: "The properties of communication."
          required: true
          schema:
            $ref: "#/definitions/emailMessageRequest"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/emailMessageResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'
  /v2/update/template:
    put:
      tags:
        - TemplateManagement
      summary: "Preview the content of the message with given parameters."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: ""
          required: true
          schema:
            $ref: "#/definitions/BodyPreview"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/PreviewResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'
  /v2/template/event_type:
    get:
      tags:
        - TemplateManagement
      summary: "Get the sending status of a specific communication."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: eventTypeUuid
          type: string
          description: "An UUID for this communication, generated by MessageHub system."
          required: true
        - in: query
          name: communicationChannel
          type: string
          description: "An UUID for this communication, generated by MessageHub system."
          required: true
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/emailMessageResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'
  /v2/template/uuid:
    get:
      tags:
        - TemplateManagement
      summary: "Get the sending status of a specific communication."
      description: ""
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: uuid
          type: string
          description: "An UUID for this communication, generated by MessageHub system."
          required: true
        - in: query
          name: communicationChannel
          type: string
          description: "An UUID for this communication, generated by MessageHub system."
          required: true
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/emailMessageResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'
  /v2/templates:
    post:
      tags:
        - TemplateManagement
      summary: "Get the sending history according to the given criterias."
      description: "`NOT RECOMMENDED TO USE`"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: "The properties of communication."
          required: true
          schema:
            $ref: "#/definitions/templateManagementRequest"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/emailMessageResponse"
        "401":
          description: '{"message":"Bad credentials"}'
        "404":
          description: '{"message":"Ressource not found"}'
        "500":
          description: '{"message":"Server error, please contact the administrator"}'
definitions:

  templateManagementRequest:
    type: object
    properties:
      communicationChannel:
        type: string
        description: "communicationChannel"
        example: "communicationChannel"
      uuid:
        type: integer
        description: "uuid"
        example: "template-email-a4a156e52e3f4bb58ddb16af181279f5"
      eventTypeUuid:
        type: string
        description: "eventTypeUuid"
        example: "eventTypeUuid"
      templateLanguageCode:
        type: string
        description: "templateLanguageCode"
        example: "templateLanguageCode"
      templateCountryCode:
        type: string
        description: "templateCountryCode"
        example: "templateCountryCode"
      startEffectiveTime:
        type: string
        description: "startEffectiveTime"
        example: "2012-11-11 12:12:12"
      lastEffectiveTime:
        type: integer
        description: "lastEffectiveTime"
        example: "2030-11-11 12:12:12"
      templateAttributes:
        type: string
        description: "templateAttributes"
        example: "templateAttributes"
  eventTypeManagementRequest:
    type: object
    properties:
      creator:
        type: string
        description: "creator"
        example: "creator"
      enableStatus:
        type: integer
        description: "enableStatus"
        example: 1
      eventName:
        type: string
        description: "eventName"
        example: "eventName"
      eventParam:
        type: string
        description: "eventParam"
        example: 1
      eventRemark:
        type: string
        description: "eventRemark"
        example: "2019-11-11 12:12:12"
      eventSource:
        type: string
        description: "eventSource"
        example: "eventSource"
      isDeleted:
        type: integer
        description: "isDeleted"
        example: 1
      notifyUrl:
        type: string
        description: "notifyUrl"
        example: "notifyUrl"
      systemId:
        type: string
        description: "systemId"
        example: "systemId"
  eventTypeManagementResponse:
    type: object
    properties:
      code:
        type: string
        description: "code"
        example: "00"
      success:
        type: boolean
        description: "success"
        example: true
      msg:
        type: string
        description: "string"
        example: "insert success"
      data:
        $ref: "#/definitions/eventTemplateManagementData"
      time:
        type: string
        description: "string"
        example: "2019-11-05T02:04:39.772+0000"

  emailMessageRequest:
    type: object
    properties:
      communicationChannel:
        type: string
        description: "communicationChannel"
        example: "Email"
      edmChannel:
        type: string
        description: "edmChannel"
        example: "decathlon"
      templateSubject:
        type: string
        description: "templateSubject"
        example: "templateSubject"
      templateContent:
        type: string
        description: "templateContent"
        example: "templateContent"
      documentName:
        type: string
        description: "documentName"
        example: "documentName"
      documentUrl:
        type: integer
        description: "documentUrl"
        example: "documentUrl"
      lastEffectiveTime:
        type: string
        description: "lastEffectiveTime"
        example: "2019-11-06 11:11:11"
      startEffectiveTime:
        type: string
        description: "startEffectiveTime"
        example: "2019-11-06 11:11:11"
      templateLanguageCode:
        type: string
        description: "chinese"
        example: "chinese"
      templateCountryCode:
        type: string
        description: "china"
        example: "china"
      templateAttributes:
        type: string
        description: "templateAttributes"
        example: "templateAttributes"
  emailMessageResponse:
    type: object
    properties:
      code:
        type: string
        description: "code"
        example: "00"
      success:
        type: boolean
        description: "success"
        example: true
      msg:
        type: string
        description: "string"
        example: "insert success"
      data:
        $ref: "#/definitions/emailMessageResponseData"
      time:
        type: string
        description: "string"
        example: "2019-11-05T02:04:39.772+0000"

  EventTypeManagementBodyPreview:
    type: object
    properties:
      enableStatus:
        type: integer
        description: "enableStatus"
        example: 1
      uuid:
        type: string
        description: "uuid"
        example: "event-type-198e6b2df21d4ea4bc597cff7cccac3c"

  BodyPreview:
    type: object
    properties:
      communicationChannel:
        type: string
        description: "communicationChannel"
        example: "Sms"
      isDeleted:
        type: integer
        description: "isDeleted"
        example: 1
      uuid:
        type: string
        description: "uuid"
        example: "template-sms-7df3293b9c81412a810b02972908a909"
  PreviewResponse:
    type: object
    properties:
      code:
        type: string
        description: "messageStatus"
        example: "00"
      success:
        type: boolean
        description: "uuid"
        example: true
      msg:
        type: string
        description: "msg"
        example: "update success"
      data:
        type: integer
        description: "data"
        example: 1
      time:
        type: string
        description: "time"
        example: "2019-11-05T03:58:40.498+0000"
  eventTemplateManagementData:
    type: object
    properties:
      uuid:
        type: string
        description: "uuid"
        example: "uuid"
      eventSource:
        type: string
        description: "eventSource"
        example: "eventSource"
      eventName:
        type: string
        description: "eventName"
        example: "eventName"
      eventParam:
        type: string
        description: "eventParam"
        example: "eventParam"
      systemId:
        type: string
        description: "systemId"
        example: "systemId"
      eventRemark:
        type: string
        description: "eventRemark"
        example: "eventRemark"
      enableStatus:
        type: integer
        description: "enableStatus"
        example: "enableStatus"
      startEffectiveTime:
        type: string
        description: "startEffectiveTime"
        example: "startEffectiveTime"
      notifyUrl:
        type: string
        description: "notifyUrl"
        example: "notifyUrl"
      creator:
        type: string
        description: "creator"
        example: "creator"
      gmtCreated:
        type: string
        description: "gmtCreated"
        example: "gmtCreated"
      modifier:
        type: string
        description: "modifier"
        example: "modifier"
      gmtModified:
        type: string
        description: "gmtModified"
        example: "gmtModified"
      isDeleted:
        type: string
        description: "isDeleted"
        example: "isDeleted"
      isDeletedVal:
        type: string
        description: "isDeletedVal"
        example: "isDeletedVal"
      enableStatusVal:
        type: string
        description: "enableStatusVal"
        example: "enableStatusVal"
  emailMessageResponseData:
    type: object
    properties:
      uuid:
        type: string
        description: "uuid"
        example: "uuid"
      eventTypeUuid:
        type: string
        description: "eventTypeUuid"
        example: "eventTypeUuid"
      edmChannel:
        type: string
        description: "officialAccount"
        example: "officialAccount"
      templateSubject:
        type: string
        description: "templateSubject"
        example: "templateSubject"
      templateContent:
        type: string
        description: "templateContent"
        example: "templateContent"
      documentName:
        type: string
        description: "documentName"
        example: "documentName"
      documentUrl:
        type: string
        description: "documentUrl"
        example: "documentUrl"
      startEffectiveTime:
        type: string
        description: "startEffectiveTime"
        example: "startEffectiveTime"
      lastEffectiveTime:
        type: string
        description: "lastEffectiveTime"
        example: "lastEffectiveTime"
      templateLanguageCode:
        type: string
        description: "cardConnection"
        example: "cardConnection"
      templateCountryCode:
        type: string
        description: "templateCountryCode"
        example: "templateCountryCode"
      communicationChannel:
        type: string
        description: "communicationChannel"
        example: "communicationChannel"
      templateAttributes:
        type: string
        description: "templateAttributes"
        example: "templateAttributes"
      creator:
        type: string
        description: "creator"
        example: "creator"
      gmtCreated:
        type: string
        description: "gmtCreated"
        example: "gmtCreated"
      modifier:
        type: string
        description: "modifier"
        example: "modifier"
      gmtModified:
        type: string
        description: "gmtModified"
        example: "gmtModified"
      isDeleted:
        type: string
        description: "isDeleted"
        example: "isDeleted"
      isDeletedVal:
        type: string
        description: "isDeletedVal"
        example: "isDeletedVal"